more work, less compile...
This is kind of a roadblock code wise, where the architecture does not
meet the right abstractions. The ABCI module needs to verify the
signature but by the point it has access to the message the signature
has been stripped off. Adding the signature everywhere would just
make the API more complicated, so I need to split it into 2. I require
more Vespene Gas (and hammock time).