ABCI bridge works
Its not perfect (response is always empty for commands so fails on cbor decode and verification)
but one can mint (using my PEM file), check a balance, and send tokens between identities.

Next step is going to properly specify each actions (and implement them) for the ledger module,
then specify the blockchain and ABCI modules properly so we can explore blocks (and implement).